refactor: split translation utils from init.py
